Move Extension to its own file / submodule
